feat: Fully replace parsing with nom
