+++ ccd_lcamv06/src/tests/single_reading_example.txt
