+++ ccd_lcamv06/src/tests/stream_example.txt
